pyproject.toml
[tool.poetry]
name = "flask-app"
version = "0.1.0"
description = "A flask app for job interview - Code Graphers"
authors = ["Hasnain hasnainnafees861@gmail.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.7.4"
flask-jwt-extended = "^4.6.0"
loguru = "^0.7.2"
bson = "^0.5.10"
motor = "^3.4.0"
flask-restx = "^1.3.0"
sentry-sdk = "^2.6.0"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
pytest = "^8.2.2"

[tool.poetry.group.dev.dependencies]
Flask = "^3.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
